Move copy_helper to utils/ dir.
