+++ gslib/utils/copy_helper.py
